requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.10
langchain-openai==0.0.8
langchain-community==0.0.13
openai==1.3.8
chromadb==0.4.24
pymupdf==1.23.8
camelot-py==0.11.0
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0
werkzeug==3.0.1
gunicorn==21.2.0
langchain-groq==0.1.0
langchain-huggingface==0.1.0
sentence-transformers==3.1.1
